[TASK] using proxy factory
Warning: pyat proxy currently only a mimic hack
