+++ src/dt4acc/interfaces/proxy_factory_interface.py
@@ -1,0 +1,8 @@
+from abc import abstractmethod, ABCMeta
+from .element_interface import ElementInterface
+
+
+class ProxyFactoryInterface(metaclass=ABCMeta):
+    @abstractmethod
+    def get(self, element_id) -> ElementInterface:
+        raise NotImplementedError
