+++ src/tests/test_nvacc.py
@@ -16,11 +16,11 @@ def test_update_quad_offset_vert():
 
 
 def test_update_quad_offset_vert():
-    update(element_id='S1MD8R', property_name='dx', value=0.5)
+    update(element_id='Q1M1D8R', property_name='dx', value=0.5)
 
-@pytest.mark.skip
+#@pytest.mark.skip
 def test_update_horizontal_steerer():
-    update(element_id='Q1M1D1R', property_name='K', value=1.2)
+    update(element_id='HS4M2D1R', property_name='K', value=1.2)
 
 
 def test_update_roll():
